Added: Add `--check-draw-over-apps-permission` flag support for `start` and `start-service` commands to check if Draw Over Apps permission has been granted
We do not always check for the permission since granting draw over apps permission will result in higher privileges for the app which it normally shouldn't have just to start activities and services. Callers who are willing to grant the permission and want to start them from background can pass the flag as an optional opt-in.

Co-authored-by: agnostic-apollo <agnosticapollo@gmail.com>
Co-authored-by: Twaik Yont <twaikyont@gmail.com>

Closes #7